Extension of platform support
I guess this would be a nice option, and given how unions and structs work, the code would not care about the changes to the number of elements. In other words, this would be left as a redundancy, and whether or not the rest of the operating system makes any use of PowerPC, MIPS or SPARC should not affect the usability of this header or the build of the project. You can consider removing them in pre-release if they turn out to be useless.